requirements.txt
autopep8==1.5.4
boto3==1.14.48
botocore==1.17.48
certifi==2020.6.20
chardet==3.0.4
docutils==0.15.2
idna==2.10
jmespath==0.10.0
mysql-connector-python==8.0.21
protobuf==3.13.0
pycodestyle==2.6.0
python-dateutil==2.8.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
urllib3==1.25.10
